Day333✅/requirements.txt
fastapi
uvicorn
requests
streamlit
python-dotenv
pydantic